streamlit/pyproject.toml
[project]
name = "stock-monitor"
version = "0.1.0"
description = "Real-time stock price monitor using Streamlit"
requires-python = ">=3.11"
dependencies = [
    "streamlit==1.32.2",
    "redis==5.0.4",
    "pandas==2.2.2",
    "python-dotenv==1.0.1",
]

[tool.uv]
package = false
